feat: refactor HeroCardImage component to use local images and simplify HeroSection
