--- src/lib/types/home/hero.ts
@@ -1,7 +1,0 @@
-import { StaticImageData } from "next/image";
-
-export type HeroSectionProp = {
-  src: StaticImageData;
-  alt: string;
-  className: string;
-};
